- fixed: use uint64_t instead of int for storing packet loss values

+++ Homer/src/Widgets/OverviewDataStreamsWidget.cpp
@@ -477,7 +477,7 @@ void OverviewDataStreamsWidget::FillRow(QTableWidget *pTable, int pRow, PacketSt
     FillCellText(pTable, pRow, 3, Int2ByteExpression(tStatValues.AvgPacketSize) + " bytes");
     FillCellText(pTable, pRow, 4, Int2ByteExpression(tStatValues.ByteCount) + " bytes");
     FillCellText(pTable, pRow, 5, Int2ByteExpression(tStatValues.PacketCount));
-    FillCellText(pTable, pRow, 6, Int2ByteExpression(tStatValues.LostPacketCount));
+    FillCellText(pTable, pRow, 6, Int2ByteExpression((int64_t)tStatValues.LostPacketCount));
     FillCellText(pTable, pRow, 8, QString(pStats->GetTransportTypeStr().c_str()));
     pTable->item(pRow, 8)->setTextAlignment(Qt::AlignCenter|Qt::AlignVCenter);
     FillCellText(pTable, pRow, 9, QString(pStats->GetNetworkTypeStr().c_str()));
